+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -499,6 +499,12 @@ Wenn diese Option aktiviert ist, werden neue Gäste unter diesen Umständen blocki
             builder.AddOrUpdate("Smartstore.AI.Prompts.ParagraphHeadingTag",
                 "Use {0} tags for the paragraph headings.",
                 "Nutze für die Überschriften der Abschnitte {0}-Tags.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.AddAlternateSitemapLinks",
+                "Add localized links",
+                "Lokalisierte Links hinzufügen",
+                "Specifies whether to add alternate links (xhtml:link) for localized page versions to the XML Sitemap.",
+                "Legt fest, ob alternative Links (xhtml:link) für lokalisierte Versionen einer Seite zur XML-Sitemap hinzugefügt werden.");
         }
     }
 }